Add the refactored visualizer module and connect all the profiler modules together.
The profiler is now easily usable and extendable as a whole. The visualizer module
contains interface which allows to further expand the supported visualizers libraries
(with some changes according to the specifics of the libraries). A wrapper module was
created, which exports simple functions that take care of the whole analysis / visualisation
connections in one place and offer simple usage for the client.

analysis.py: the wrapper module which handles the analysis and visualisation using simple functions
complexity.py: updated to reflect the changes in modules structure
interface.py: the unified interface for visualizer modules
bokeh_visualizer.py: the bokeh visualisation library, first of the visualizers
visualizer_exceptions.py: the visualization-specific exception hierarchy
tools.py, utils.py: the private/public helper function and tools
renames: updated the structure a bit