Support more smart lookup of the profile for add
cli: add massing of profile name for the perun add;
rest: minor refactorings.

(*) Add feature: Profile lookup for Add is more smart, i.e. it tries
to check first if the file exists, or if it is in pending jobs and then
walk through the directory tree and tries to find anything matching
the string and asks for confirmations.