Add tracking of the origin of the generated profile
add: add origin to profile so it can be compared before adding profile;
add: by default remove the profile that is added to perun;
pcs: add helper function for getting head of the wrapped vcs;
store: instead of storing path, store just the filename (***);
tests: fix tests according to the current implementations.

(*) Add feature: Add tracking of the origin of the file---when the profile
is generated, one has to know to which version it originally corresponded,
so you cannot store it for incorrect minor version. Before it is added to perun,
the origin is however removed, because we want to minimize possible diffs
of profile so we can exploit the compact format of zlib.

(**) Add feature: Remove profile when add

(***) Having full (or relative) path in index is strange actually. Since once collected,
profile names are generated with their creation time and so this will be unique some
how. Moreover, since they can correspond to different minor versions this adds
another layer, so path is not needed and eases some pain.