Add simple allocation queries (#20)
This pull request introduces the alloclist textual interpretation of
memory profiles. It is a set of frequent queries over the allocation data,
output in a list format. Optionally one can limit the list to the specific
function, to certain interval of the timestamps.

Authored by: Radim Podola 
Reviewed and Refactored by: Tomas Fiedor