Add paging for log and status
commands: add paging to log and status;
log: pager decorator for paging of output;
cli: add option to enable/disable paging when run from cli (*).

(*) Currently the paging is meant only for times when perun is run from CLI
directly. When it is run from tests or from scripts, it is disable by default. This
behaviour can be suppressed by setting log.SUPPRESS_PAGING to false, then
it will be paged all the time.