Enhance the performance of Perun
commands: add else branch for checking the minor version validity;
store: guard the hashlib by demand import;
utils: rewrite the get_supported_module_names to table lookup (*);
cli: add lazy init of unit commands (**);
test_utils: add sanity checks for registered modules.

(*) Note this used to be dynamic, but hardcoded table yield a great gain
in performance. This is now guarded by sanity checks in test_utils, since
registering is rare occurence and hence we do not have to do a dynamic
lookup in my opinion.

(**) This prevents the dynamic registered even though we do not need
to do the initialization as the part of the click is not called at all.