Fix issue with profile file lookup
commands: fix issue with uncreated index (*);
utils: change assert to error with fail;
test_cli: add tests for absolute and relative paths.

(*) This fixes issue, when one wants to use the absolute or relative path
lookup of profile, but the index was not yet created and it then failed during
index lookup and the lookup never occured.