Fix two issues for Continuous Integration
commands: add sorting of the untracked files (*);
test_query: copy the generator of profiles for query (**).

(*) This caused problems in the CI, where it assumed different ordering
of the profiles and suddenly (probably after updated screwed things
over).

(**) This caused problems in CI tests as well, where it again reversed
the order, so the profiles were screwed.