Add streaming of vcs history and output
check: make precollect silent (output to devnull);
runner: make two instances of run function (one with history and one without);
log: add autoflush to history object (to stream and not print when ready);
log: flush history only when the input stdout is readable (*).

(*) This is needed in order to be compatible with silencing of precollect.