Create PyQt-Fit local port.
The PyQt-Fit package is no longer maintained but our nonparametric
kernel regression module depends on it. So far, we have been using
a custom wheel and constrained dependencies to hotfix the issue.

This commit creates a custom PyQt-Fit port as a local package that
can be easily imported and maintained if needed.