Finish the PyQt-Fit and bkcharts removal.
This commit finalizes the removal of custom PyQt-Fit and bkcharts
wheel dependencies in Perun. PyQt-Fit is now locally ported and
bkcharts has been substituted with holoviews.

This commit also removes some upper limits on dependency versions
(most notably pandas, numpy and bokeh) that were imposed because
of the custom wheels.