Remove the Codecov token from workflow.
Repository secrets are not working properly for GH Actions triggered
by PRs from forked repositories. Although we might need to retry the
Codecov upload from time to time due to an upload error, it is still
the better solution than to hard-code the token or miss a coverage
upload from new contributors.