build: Port to Meson build system
Powerful build system for building potential Python extensions and in
general any C/C++/<your-favourite-language> code.

Straightforward Python-like syntax, integration with editable
environments using automatic rebuilds, source/wheel distribution, build
introspection, reproducible builds and more!

The Makefile is still present as all interaction with Meson should be
done via pip/pypa-build and getting all the options correct is just too
much.

All references to setuptools are gone and instead there are new
dependencies on Meson, meson-python, ninja and patchelf. Meson being the
build system, meson-python the project linking the Python ecosystem with
Meson, ninja the build tool and patchelf a required dependency of
meson-python. All are easily accessible across platforms.

The best part of all this is that the only visible change is that you
need the new dependencies. The rest of the development process should be
the same as it was before the port.

Closes https://github.com/Perfexionists/perun/issues/125