Profiles are no longer overwritten by-default
The naming and saving of profiles has been unified such that it
is now done only in one place in the code. The profile stem
(the name without a suffix) is now being extended with a copy
number, e.g., 'custom_profile(N).perf' in case a profile with
the same name already exists in the directory.